file restructure for client folder
