+++ client/components/auth-form.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import {connect} from 'react-redux'
 import PropTypes from 'prop-types'
-import {auth} from '../store'
+import {auth} from '../store/actions/users'
 
 /**
  * COMPONENT
