+++ client/components/navbar.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
 import {Link} from 'react-router-dom'
-import {logout} from '../store'
+import {logout} from '../store/actions/users'
 
 const Navbar = ({handleClick, isLoggedIn}) => (
   <div>
