+++ client/store/reducers/index.js
@@ -1,0 +1,4 @@
+import user from './user'
+import {combineReducers} from 'redux'
+
+export default combineReducers({user})
